+++ mobile/build.gradle.kts
@@ -83,12 +83,12 @@ dependencies {
     implementation(projects.feature.subscription.dialog.impl)
     implementation(projects.feature.subscription.list.api)
     implementation(projects.feature.subscription.list.impl)
-    implementation(projects.feature.transaction.csvimport.api)
-    implementation(projects.feature.transaction.csvimport.impl)
     implementation(projects.feature.transaction.detail.api)
     implementation(projects.feature.transaction.detail.impl)
     implementation(projects.feature.transaction.editor.api)
     implementation(projects.feature.transaction.editor.impl)
+    implementation(projects.feature.transaction.importer.api)
+    implementation(projects.feature.transaction.importer.impl)
     implementation(projects.feature.transaction.overview.impl)
     implementation(projects.feature.transfer.api)
     implementation(projects.feature.transfer.impl)
