+++ mobile/src/main/kotlin/ru/resodostudios/cashsense/ui/CsApp.kt
@@ -73,11 +73,11 @@ import ru.resodostudios.cashsense.feature.settings.impl.navigation.settingsEntry
 import ru.resodostudios.cashsense.feature.subscription.dialog.api.navigateToSubscriptionDialog
 import ru.resodostudios.cashsense.feature.subscription.dialog.impl.navigation.subscriptionDialogEntry
 import ru.resodostudios.cashsense.feature.subscription.list.impl.navigation.subscriptionsEntry
-import ru.resodostudios.cashsense.feature.transaction.csvimport.api.ImportNavKey
-import ru.resodostudios.cashsense.feature.transaction.csvimport.impl.navigation.importDialogEntry
 import ru.resodostudios.cashsense.feature.transaction.detail.impl.navigation.transactionEntry
 import ru.resodostudios.cashsense.feature.transaction.editor.api.TransactionEditorNavKey
 import ru.resodostudios.cashsense.feature.transaction.editor.impl.navigation.transactionEditorEntry
+import ru.resodostudios.cashsense.feature.transaction.importer.api.ImportNavKey
+import ru.resodostudios.cashsense.feature.transaction.importer.impl.navigation.importDialogEntry
 import ru.resodostudios.cashsense.feature.transaction.overview.impl.navigation.transactionOverviewEntry
 import ru.resodostudios.cashsense.feature.transfer.impl.navigation.transferDialogEntry
 import ru.resodostudios.cashsense.feature.wallet.detail.api.WalletNavKey
