+++ settings.gradle.kts
@@ -61,12 +61,12 @@ include(":feature:subscription:dialog:api")
 include(":feature:subscription:dialog:impl")
 include(":feature:subscription:list:api")
 include(":feature:subscription:list:impl")
-include(":feature:transaction:csvimport:api")
-include(":feature:transaction:csvimport:impl")
 include(":feature:transaction:detail:api")
 include(":feature:transaction:detail:impl")
 include(":feature:transaction:editor:api")
 include(":feature:transaction:editor:impl")
+include(":feature:transaction:importer:api")
+include(":feature:transaction:importer:impl")
 include(":feature:transaction:overview:api")
 include(":feature:transaction:overview:impl")
 include(":feature:transfer:api")
